+++ src/app/user.service.ts
@@ -1,0 +1,17 @@
+import { HttpClient } from '@angular/common/http';
+import { Injectable } from '@angular/core';
+import { UserModel } from './models/user.model';
+import { environment } from 'src/environments/environment';
+import { Observable } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class UserService {
+  constructor(private http: HttpClient) {}
+
+  register(login: string, password: string, birthYear: number): Observable<UserModel> {
+    const body = { login, password, birthYear };
+    return this.http.post<UserModel>(`${environment.base_url}/api/users`, body);
+  }
+}